[v1.0] Step 3: 'JPA based Extension Model' of the WebShop Extension Scenario
This commit contains the sources of the 'JPA based Extension Model'
component of the WebShop Extension Scenario. This component is part of
the WebShop Extension Application ('Customer Reviews') which is running
as demo application on SAP HANA Cloud
https://webreviewsespmhana.hana.ondemand.com/espm-ui-reviews-web/

The corresponding WebShop Extension Scenario description on SCN
http://www.sdn.sap.com/irj/scn/index?rid=/library/uuid/8016f858-369d-3010-1f9c-9402c4ecad3fis
is a step by step tutorial for any developer. 
In Chapter 3 it is described how 'JPA based Extension Model' part is
build up and how it runs.

Used technologies SAPUI5, OData, SAP HANA Cloud Connectivity